Releasing applications to prod (#49)
* Update develop (#46)

* Update index.jsx

* owo'

* changes

* Added folder Login in 'src/sections' along with index.jsx and corresponding empty Login.module.scss

* Updated Login page with more styling.

* owo

* Created border boxes in login form and adjusted sign in button to look clean

* Dashboard main (#32)

* dashboard implementation (status not completed yet)

* margin fix

* application status implementation

* Added file dropper implementation (#29)

* Added file dropper implementation

* Removed testing code

* added requested changes

* Got rid of testing code

* Added props for supported file extensions and file input name attribute;
Improved error logic so that it will display to the user that they dropped in an invalid file;
Added comments to explain the props

* owo

* Responsive login

* owo final polish

* uwu

* clear auto complete thingy

* init application page

* Added featureflags and prashant changes

* fix build issues (TODO: ssr level routing)

* Added proper route filter + 404

* update date

* Added sponsor stuff

* fixed sponsor hydration styles

* removed sponsor type header

* Update data.js

* Forgot Password Page (#36)

* Added forgot password page

* text lineheight

* Checkbox + Layout for some Application sections (#35)

* Added checkbox and added layout for experience section and below

* Updated checkbox with better logic, focus css, and styling

* using useRef instead of createRef

* Added form controls for experience, MLH, and consent sections;
Added Required component


* Update gatsby-config.js

* Update index.jsx

* added virtual (#37)

* Finished apps (#38)

* touched up guards and error handling

* lint

* Update FormContext.jsx

* bug fixes

* fix: change factions to routes

* fix: More routing fixes

* Update index.jsx

* Update majors.js

* Added forgot password

* fix window check

* owo

* Update schools.js

* Update gatsby-config.js

* Update index.jsx

* Update schools.js

* Added rsvp ability

* Update index.jsx

* Fixed status error

* more state changes

* Added discord stuff

* Update index.jsx

* task: Added rsvp feature flag

* Update gatsby-config.js

* task: fix rsvp toggle

* Update gatsby-config.js

* fix: Added netlify plugin to gatsby

* welp

* welp

* owo

* save me

* uwu

* Added pinnguaq challenge

* Added The Hub logo with conditional rendering (#39)

* task: Added schedule

* task: Touched up schedule linking

* Fixed date/time bug for schedule

* removed log

* Fixed default val

* Update Events.module.scss

* Updated Splash for HTV 6 (#41)

* Change metadata in preparation for HTV 6

* Edit Splash for pre-registration HTV

* Updated dates for HTV

* Commented out sponsors card (#42)

* Commented out sponsors card

* Update package.json


* Update about date to Winter 2022 (#45)

* Update about date to Winter 2022

* Update package.json


* Updated HTV date in FAQs (#43)

* Update HTV date

* Update package.json

* Update dates based on gatsby-config module

* Commented out Pinnguaq/The Hub faction (#44)

* Removed Pinnguaq/The Hub faction

Commented out data for Pinnguaq/The Hub faction...may have them back?

* Update version

* Set grid column dynamically using css variables



* Release sponsors (#47)

* Release sponsors

* Add feature flag for sponsors

* Added featureflag implementation

* owo sponsors

* owo forms stuff

* fix form misorder

+++ gatsby-node.js
@@ -4,17 +4,23 @@ const isProd = process.env.NODE_ENV === `production`;
 
 exports.onCreatePage = ({ page, actions }) => {
   // Remove pages protected by featureFlag
-  if (!siteMetadata.featureFlags.open) {
+  if (isProd && !siteMetadata.featureFlags.open) {
     if (
-      ['/application', '/dashboard', '/register', '/login'].includes(
+      ['/application', '/dashboard', '/activate', '/forgot', '/reset_password', '/register', '/login'].includes(
         replacePath(page.path),
       )
     ) {
       actions.deletePage(page);
     }
   }
+
+  if (isProd && !siteMetadata.featureFlags.schedule) {
+    if (['/schedule'].includes(replacePath(page.path))) {
+      actions.deletePage(page);
+    }
+  }
   
-  if (!siteMetadata.featureFlags.discord) {
+  if (isProd && !siteMetadata.featureFlags.discord) {
     if (['/discord'].includes(replacePath(page.path))) {
       actions.deletePage(page);
     }
